+++ lib/src/constant/const.dart
@@ -1,3 +1,4 @@
 class Const {
   static const String localIp = "127.0.0.1";
+  static const String defaultIp = "0.0.0.0";
 }