+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.9
+- Headers & Token implemented
+- Able to fetch client info like (ip, port etc.)
+
 ## 1.0.8
 - Execute dart format .
 - Add gif